MYR-75 Decompose export route + extract shared error envelope
Addresses Claude Review on PR #259:

Blocking: route.ts at 510 lines exceeded the 200-line cap.
- Extract 10 Export* interfaces to types.ts (155 lines, declarative).
- Extract entity mapper functions (mapProfileToExport,
  mapAccountToExport, mapSettingsToExport, mapVehicleToExport,
  mapDriveToExport, mapInviteToExport, mapAuditLogToExport) to
  mappers.ts (295 lines, pure transformations).
- route.ts is now 187 lines — auth → transaction → map → return.

Blocking: errorEnvelope was copy-pasted between the DELETE handler
and the export handler.
- Extract to src/lib/api-errors.ts (22 lines).
- Update both routes to import from the shared module.
- DELETE handler is now 84 lines (was ~99).

Suggestion #3 (non-blocking, taken): use auditEntry.timestamp.toISOString()
for ExportArchive.exportedAt so it equals the canonical audit moment
exactly — no millisecond drift between two `new Date()` calls.

All 29 tests in __tests__/app/api/users/me/ pass; full vitest suite
green; lint clean (9 pre-existing warnings unrelated); build clean.
